+++ handlers/update-supporter-plus-amount/src/updateSupporterPlusAmount.ts
@@ -1,7 +1,7 @@
 import { getSingleOrThrow } from '@modules/arrayFunctions';
 import { ValidationError } from '@modules/errors';
-import type { Currency } from '@modules/internationalisation/currency';
-import { isSupportedCurrency } from '@modules/internationalisation/currency';
+import type { IsoCurrency } from '@modules/internationalisation/isoCurrency';
+import { isSupportedCurrency } from '@modules/internationalisation/isoCurrency';
 import type { Logger } from '@modules/logger';
 import { getIfDefined } from '@modules/nullAndUndefined';
 import { prettyPrint } from '@modules/prettyPrint';
@@ -132,7 +132,7 @@ export const getSupporterPlusData = (
 
 const validateNewAmount = (
 	newAmount: number,
-	currency: Currency,
+	currency: IsoCurrency,
 	billingPeriod: ProductBillingPeriod<'SupporterPlus'>,
 ): void => {
 	const amountBand = supporterPlusAmountBands[currency][billingPeriod];
